Added style merge test
While implementing it, I realized that pulling values out of a hashset
didn't guarantee order, and for these tests to pass, order needed to be
guaranteed. Wrote a specialized iterator to do this operation, as I
don't currently rely on itertools and it seemed like it wasn't going to
be hard to wite a specialized version that could assume the input
iterators are sorted.

So, Classes now sorts itself during construction, and merging will no
longer uses an intermediate allocation to do the unique operation.