Moving code to gooey
These don't have any tie in to the core functionality of the library. I had moved all the previous components out, it seemed odd for these to still be in stylecs.

This moves the palette dependency up the chain as well, meaning this library is a much lighter weight inclusion if someone else could use it.